requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-modeladmin-reorder==0.3.1
django-simple-history==3.7.0
django-suit==0.2.28
django-unfold==0.39.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
drf-yasg==1.21.7
executing==2.1.0
Faker==30.8.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipython==8.27.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pluggy==1.5.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
resend==2.4.0
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.8.2
